Optimize pages with server components + Convex preloadQuery
Convert client pages to server components where possible, and migrate
fetchQuery to preloadQuery for SSR with seamless real-time hydration.

- Remove unnecessary "use client" from sign-in page
- Convert payment merci/cancelled to server (searchParams prop)
- Post detail: preloadQuery + usePreloadedQuery (no loading spinner)
- Profile layout: server component with preloadQuery + UserProfileShell
- Profile/gallery pages: consume layout context (eliminate redundant queries)
- Income page: server auth guard + redirect, preloaded earnings data
