+++ app/(app-pages)/payment/cancelled/page.tsx
@@ -1,13 +1,12 @@
-"use client"
-
 import { CircleX } from "lucide-react"
 import Link from "next/link"
-import { useSearchParams } from "next/navigation"
+import { use } from "react"
 import { buttonVariants } from "@/components/ui/button"
 
-const CancelledPage = () => {
-  const search = useSearchParams()
-  const username = search.get("username")
+const CancelledPage = (props: {
+  searchParams: Promise<{ username?: string }>
+}) => {
+  const { username } = use(props.searchParams)
   return (
     <main className="border-muted flex h-full min-h-screen w-full flex-col items-center justify-center border-r border-l">
       <div className="flex flex-col items-center gap-10">
