+++ components/domains/users/index.ts
@@ -6,4 +6,5 @@ export { UserPosts } from "./user-posts"
 export { UserProfileHeader } from "./user-profile-header"
 export { UserProfileHero } from "./user-profile-hero"
 export { UserProfileTabs } from "./user-profile-tabs"
+export { UserProfileShell, useUserProfileContext } from "./user-profile-shell"
 export { UserReportButton } from "./user-report-button"
